requirements.txt
streamlit
google-cloud-vision
openai
pandas
firebase-admin
python-dotenv
requests